Fixed Authroute and Signup

+++ src/App.js
@@ -5,7 +5,7 @@ import jwtDecode from 'jwt-decode'
 import axios from 'axios'
 
 //Components
-import AuthRoute from './util/AuthRoute'
+// import AuthRoute from './util/AuthRoute'
 import Navbar from './components/Navbar'
 import Footer from './components/Footer'
 
@@ -36,19 +36,25 @@ export default class App extends Component {
   constructor(props) {
     super(props)
     this.state = {
+      handle: '',
       email: '',
       password: '',
-      user: {
-        handle: '',
-        userId: '',
-        email: '',
-        createdAt: '',
-        role: '',
-        store_location: ''
-      }
+      confirmPassword: '',
+      store_location: '',
+      role: ''
+
+      // user: {
+      //   handle: '',
+      //   userId: '',
+      //   email: '',
+      //   createdAt: '',
+      //   role: '',
+      //   store_location: ''
+      // }
     }
     this.handleInput = this.handleInput.bind(this)
-    this.handleSubmit = this.handleSubmit.bind(this)
+    this.handleSubmitLogin = this.handleSubmitLogin.bind(this)
+    this.handleSubmitSignup = this.handleSubmitSignup.bind(this)
   }
 
   //Input handler
@@ -59,8 +65,8 @@ export default class App extends Component {
     })
   }
 
-  //Submit data handler calling Auth Service
-  handleSubmit = e => {
+  //Submit data handler for Login
+  handleSubmitLogin = e => {
     if (e) e.preventDefault()
     const userData = {
       email: this.state.email,
@@ -68,6 +74,29 @@ export default class App extends Component {
     }
     axios
       .post(process.env.REACT_APP_API_URL + '/login', userData)
+      .then(res => {
+        localStorage.setItem('FBIdToken', `Bearer ${res.data.token}`)
+        window.location.href = '/profile'
+        return
+      })
+      .catch(err => {
+        return console.error(err)
+      })
+  }
+
+  //Submit data handler for Signup
+  handleSubmitSignup = e => {
+    if (e) e.preventDefault()
+    const userData = {
+      handle: this.state.handle,
+      email: this.state.email,
+      password: this.state.password,
+      confirmPassword: this.state.confirmPassword,
+      store_location: this.state.store_location,
+      role: this.state.role
+    }
+    axios
+      .post(process.env.REACT_APP_API_URL + '/signup', userData)
       .then(res => {
         localStorage.setItem('FBIdToken', `Bearer ${res.data.token}`)
         window.location.href = '/profile'
@@ -92,19 +121,34 @@ export default class App extends Component {
                   {...props}
                   email={this.state.email}
                   password={this.state.password}
-                  submitHandler={this.handleSubmit}
+                  submitHandler={this.handleSubmitLogin}
                   inputHandler={this.handleInput}
                 />
               )}
             />
-            <Route path="/signup" component={Signup} />
+            <Route
+              path="/signup"
+              render={props => (
+                <Signup
+                  {...props}
+                  handle={this.state.handle}
+                  email={this.state.email}
+                  password={this.state.password}
+                  confirmPassword={this.state.confirmPassword}
+                  store_location={this.state.store_location}
+                  role={this.state.role}
+                  submitHandler={this.handleSubmitSignup}
+                  inputHandler={this.handleInput}
+                />
+              )}
+            />
             <Route
               path="/profile"
               render={props => <Profile {...props} user_email={this.state.email} authenticated={authenticated} />}
             />
             <Route path="/thankyou" component={ThankYou} />
             <Route path="/about" component={About} />
-            <Route path="/confirm/:id" component={CustomerImage} authenticated={authenticated} />
+            <Route path="/confirm/:id" component={CustomerImage} />
           </Switch>
         </Router>
         <Footer />
